+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Adapt readFragPipeInfo to handle different FragPipe versions
 * Add filtering function for FragPipe
+* Add option to use sample weights for limma
 
 # einprot 0.6.9